#272 ReviewAndLiveComponent. Upgrade animation (remove checked). Added animation in review
