+++ src/app/play/brick/live.component.ts
@@ -11,124 +11,9 @@ import { CompComponent } from './comp/comp.component';
 import { QuestionComponent } from './question.component';
 import { Router, ActivatedRoute } from '@angular/router';
 import { AuthService } from '../../auth/auth.service';
-import * as mojs from 'mo-js';
 // Important jquery may not look like it is being used but it really is
 import * as $ from 'jquery';
-
-function isIOSSafari() {
-    let userAgent;
-    userAgent = window.navigator.userAgent;
-    return userAgent.match(/iPad/i) || userAgent.match(/iPhone/i);
-}
-
-function isTouch() {
-    let isIETouch;
-    isIETouch = navigator.maxTouchPoints > 0 || navigator.msMaxTouchPoints > 0;
-    return [].indexOf.call(window, 'ontouchstart') >= 0 || isIETouch;
-}
-
-// taken from mo.js demos
-const isIOS = isIOSSafari(),
-    clickHandler = isIOS || isTouch() ? 'touchstart' : 'click';
-
-function extend( a, b ) {
-    for (const key in b ) {
-        if (b.hasOwnProperty( key ) ) {
-            a[key] = b[key];
-        }
-    }
-    return a;
-}
-
-function Animocon(el, options) {
-    this.el = el;
-    this.options = extend( {}, this.options );
-    extend( this.options, options );
-
-    this.checked = false;
-
-    this.timeline = new mojs.Timeline();
-
-    for (let i = 0, len = this.options.tweens.length; i < len; ++i) {
-        this.timeline.add(this.options.tweens[i]);
-    }
-
-    const self = this;
-    this.el.addEventListener(clickHandler, function() {
-        if (self.checked) {
-            self.options.onUnCheck();
-        } else {
-            self.options.onCheck();
-            self.timeline.replay();
-        }
-        self.checked = !self.checked;
-    });
-}
-
-Animocon.prototype.options = {
-    tweens : [
-        new mojs.Burst({})
-    ],
-    onCheck : function() { return false; },
-    onUnCheck : function() { return false; }
-};
-
-// Animocon animation function
-function init(buttons) {
-    buttons.forEach(el1 => {
-        const el1span = el1.querySelector('mat-icon');
-        const notUsedObj = new Animocon(el1, {
-            tweens : [
-                // burst animation
-                new mojs.Burst({
-                    parent: el1,
-                    radius: {30: 90},
-                    count: 6,
-                    children : {
-                        fill: '#0076b4',
-                        opacity: 0.6,
-                        radius: 15,
-                        duration: 1700,
-                        easing: mojs.easing.bezier(0.1, 1, 0.3, 1)
-                    }
-                }),
-                // ring animation
-                new mojs.Shape({
-                    parent: 		el1,
-                    type: 			'circle',
-                    radius: 		{0: 60},
-                    fill: 			'transparent',
-                    stroke: 		'#0076b4',
-                    strokeWidth: {20: 0},
-                    opacity: 		0.6,
-                    duration: 	700,
-                    easing: 		mojs.easing.sin.out
-                }),
-                // icon scale animation
-                new mojs.Tween({
-                    // This is longer than the slide transition to next question
-                    // But it seems to work ok at the moment
-                    duration : 1200,
-                    onUpdate: function(progress) {
-                        if (progress > 0.3) {
-                            const elasticOutProgress = mojs.easing.elastic.out(1.43 * progress - 0.43);
-                            el1span.style.WebkitTransform = el1span.style.transform =
-                                'scale3d(' + elasticOutProgress + ',' + elasticOutProgress + ',1)';
-                        } else {
-                            el1span.style.WebkitTransform = el1span.style.transform = 'scale3d(0,0,1)';
-                        }
-                    }
-                })
-            ],
-            onCheck : function() {
-                el1.style.color = '#ffffff';
-            },
-            onUnCheck : function() {
-                el1.style.color = '#0076b4';
-            }
-        });
-    });
-}
+import { animateButtons } from 'src/app/animocon/button';
 
 @Component({
     selector: 'live',
@@ -171,7 +56,7 @@ export class LiveComponent implements OnInit {
         // Poll to check for button elements with icobutton class
         setTimeout(function() {
             const items = [].slice.call(document.querySelectorAll('button.icobutton'));
-            init(items);
+            animateButtons(items);
         }, 500);
     }
 
@@ -215,8 +100,10 @@ export class LiveComponent implements OnInit {
             };
             console.log(`score is ${score} out of ${maxScore}, which is ${score * 100 / maxScore}%`);
             this.bricks.currentBrickAttempt = ba;
-            this.router.navigate(['../provisionalScore'], { relativeTo: this.route });
-        })
+            setTimeout(() => {
+                this.router.navigate(['../provisionalScore'], { relativeTo: this.route });
+            }, 500);
+        });
     }
 
 }
