+++ src/app/play/brick/review.component.html
@@ -1,5 +1,5 @@
 <timer [timer]="timer" [isRed]="true"></timer>
-<mat-horizontal-stepper>
+<mat-horizontal-stepper #stepper>
     <ng-template matStepperIcon="number" let-index="index">
         <ng-container *ngIf="index < 20">
             <span [ngClass]="'icon-raw-blue-' + (index + 1)"></span>
@@ -15,11 +15,11 @@
             <div fxFlex="1 0 0"></div>
             <ng-container *ngIf="index === (brick | async)?.questions.length - 1; else  next_button">
                 <p class="begin-label">Finish Brick Attempt</p>
-                <a mat-fab class="fab-bottom" (click)="finishBrick()"><mat-icon>play_arrow</mat-icon></a>
+                <button mat-fab class="fab-bottom icobutton" (click)="finishBrick()"><mat-icon>play_arrow</mat-icon></button>
             </ng-container>
             <ng-template #next_button>
                 <p class="begin-label">Next - Don't panic, you can always come back</p>
-                <button class="fab-bottom" mat-fab mat-button matStepperNext><mat-icon>play_arrow</mat-icon></button>
+                <button class="fab-bottom icobutton" mat-fab mat-button (click)="goForward(stepper)"><mat-icon>play_arrow</mat-icon></button>
             </ng-template>
         </div>
     </mat-step>
